+++ IM_calculation/IM/im_calculation.py
@@ -1,91 +1,51 @@
 import csv
-from datetime import datetime
-import getpass
 import glob
 import os
 import re
 import sys
+from datetime import datetime
+from multiprocessing.pool import Pool
+from collections import ChainMap
+from typing import List
 
 import numpy as np
-from multiprocessing.pool import Pool
 import pandas as pd
 
+from qcore import timeseries, constants
+from qcore.constants import Components
+from qcore.im import order_im_cols_df
+
+from IM_calculation.IM import read_waveform, intensity_measures
 from IM_calculation.Advanced_IM import advanced_IM_factory
-from IM_calculation.IM import read_waveform
-from IM_calculation.IM import intensity_measures
-from qcore import constants
-from qcore import timeseries
+from IM_calculation.IM.computeFAS import get_fourier_spectrum
 
 
 G = 981.0
-IMS = ["PGA", "PGV", "CAV", "AI", "Ds575", "Ds595", "MMI", "pSA"]
+DEFAULT_IMS = ("PGA", "PGV", "CAV", "AI", "Ds575", "Ds595", "MMI", "pSA")
+ALL_IMS = ("PGA", "PGV", "CAV", "AI", "Ds575", "Ds595", "MMI", "pSA", "FAS")
 
-EXT_PERIOD = np.logspace(start=np.log10(0.01), stop=np.log10(10.0), num=100, base=10)
-BSC_PERIOD = [
-    0.02,
-    0.05,
-    0.1,
-    0.2,
-    0.3,
-    0.4,
-    0.5,
-    0.75,
-    1.0,
-    2.0,
-    3.0,
-    4.0,
-    5.0,
-    7.5,
-    10.0,
-]
+MULTI_VALUE_IMS = ("pSA", "FAS")
 
-COMPONENTS = ["090", "000", "ver", "geom"]
+FAS_FREQUENCY = np.logspace(-1, 2, num=100, base=10.0)
+
 
 FILE_TYPE_DICT = {"a": "ascii", "b": "binary"}
 META_TYPE_DICT = {"s": "simulated", "o": "observed", "u": "unknown"}
 RUNNAME_DEFAULT = "all_station_ims"
 
-OUTPUT_PATH = os.path.join("/home", getpass.getuser())
 OUTPUT_SUBFOLDER = "stations"
 
 MEM_PER_CORE = 7.5e8
 MEM_FACTOR = 4
 
 
-def add_if_not_exist(component_list, component):
-    if not component in component_list:
-        component_list.append(component)
-    return component_list
-
-
-def convert_str_comp(arg_comps):
-    """
-    convert arg comps to str_comps_for integer_convestion in read_waveform & str comps for writing result
-    :param arg_comps: user input a list of comp(s)
-    :return: two lists of str comps
-    """
-    # comps = ["000", "geom",'ver']
-    if "geom" in arg_comps:
-        # ['000', 'ver', '090', 'geom']
-        str_comps = add_if_not_exist(arg_comps, "090")
-        str_comps = add_if_not_exist(str_comps, "000")
-        # for integer convention, str_comps shouldn't include geom as waveform has max 3 components
-        str_comps.remove("geom")
-        # for writing result, make a copy of the str_comps for int convention, and shift geom to the end
-        str_comps_for_writing = str_comps[:]
-        str_comps_for_writing.append("geom")
-        return str_comps, str_comps_for_writing
-    else:
-        return arg_comps, arg_comps
-
-
-def array_to_dict(value, str_comps, im, arg_comps):
+def array_to_dict(value, comps_to_calc, im, comps_to_store):
     """
     convert a numpy arrary that contains calculated im values to a dict {comp: value}
     :param value: calculated intensity measure for a waveform
-    :param str_comps: a list of components converted from user input
+    :param comps_to_calc: a list of components converted from user input
     :param im:
-    :param arg_comps:user input list of components
+    :param comps_to_store:user input list of components
     :return: a dict {comp: value}
     """
     value_dict = {}
@@ -93,20 +53,20 @@ def array_to_dict(value, str_comps, im, arg_comps):
     # [0, 2]          [0, 1]                  [0, 1, 2]
     for i in range(value.shape[-1]):
         # pSA returns a 2D array
-        if im == "pSA":
-            value_dict[str_comps[i]] = value[:, i]
+        if im in MULTI_VALUE_IMS:
+            value_dict[comps_to_calc[i].str_value] = value[:, i]
         else:
-            value_dict[str_comps[i]] = value[i]
+            value_dict[comps_to_calc[i].str_value] = value[i]
     # In this case, if geom in str_comps,
     # it's guaranteed that 090 and 000 will be present in value_dict
-    if "geom" in str_comps:
-        value_dict["geom"] = intensity_measures.get_geom(
-            value_dict["090"], value_dict["000"]
+    if Components.cgeom in comps_to_store:
+        value_dict[Components.cgeom.str_value] = intensity_measures.get_geom(
+            value_dict[Components.c090.str_value], value_dict[Components.c000.str_value]
         )
-        # then we pop unwanted keys from value_dict
-        for k in str_comps:
-            if k not in arg_comps:
-                del value_dict[k]
+    # then we pop unwanted keys from value_dict
+    for k in comps_to_calc:
+        if k not in comps_to_store and k.str_value in value_dict:
+            del value_dict[k.str_value]
     return value_dict
 
 
@@ -134,15 +94,35 @@ def compute_adv_measure(waveform, advanced_im_config, output_dir):
         )
 
 
-def compute_measure_single(value_tuple):
+def calculate_rotd(spectral_displacements, comps_to_store: List[Components]):
+    """
+    Calculates rotd for given spectral displacements
+    :param spectral_displacements: An array with shape [periods.size, nt, 2] where nt is the number of timesteps in the original waveform
+    :param comps_to_store: A list of components to store
+    :return: A dictionary with the comps_to_store as keys, and 1d arrays of shape [periods.size] containing the rotd values
+    """
+    rotd = intensity_measures.calc_rotd(spectral_displacements)
+    value_dict = {}
+    if Components.crotd50 in comps_to_store:
+        value_dict[Components.crotd50.str_value] = np.median(rotd, axis=1)
+    if Components.crotd100 in comps_to_store:
+        value_dict[Components.crotd100.str_value] = np.max(rotd, axis=1)
+    if Components.crotd100_50 in comps_to_store:
+        value_dict[Components.crotd100_50.str_value] = np.max(rotd, axis=1) / np.median(
+            rotd, axis=1
+        )
+    return value_dict
+
+
+def compute_measure_single(
+    waveform, ims, comps_to_store, im_options, comps_to_calculate
+):
     """
     Compute measures for a single station
     :param: a tuple consisting 5 params: waveform, ims, comp, period, str_comps
     waveform: a single tuple that contains (waveform_acc,waveform_vel)
     :return: {result[station_name]: {[im]: value or (period,value}}
     """
-    waveform, ims, comps, period, str_comps = value_tuple
-    result = {}
     waveform_acc, waveform_vel = waveform
     DT = waveform_acc.DT
     times = waveform_acc.times
@@ -157,47 +137,81 @@ def compute_measure_single(value_tuple):
 
     station_name = waveform_acc.station_name
 
-    result[station_name] = {}
+    result = {(station_name, comp.str_value): {} for comp in comps_to_store}
 
-    for im in ims:
-        if im == "PGV":
-            value = intensity_measures.get_max_nd(velocities)
-
-        if im == "PGA":
-            value = intensity_measures.get_max_nd(accelerations)
-
-        if im == "pSA":
-            value = intensity_measures.get_spectral_acceleration_nd(
-                accelerations, period, waveform_acc.NT, DT
+    def process_single_value_im(im, func, *args, **kwargs):
+        if im in ims:
+            value = func(*args, **kwargs)
+            values_to_store = array_to_dict(
+                value, comps_to_calculate, im, comps_to_store
             )
+            for comp in comps_to_store:
+                if comp.str_value in values_to_store:
+                    result[(station_name, comp.str_value)][im] = values_to_store[
+                        comp.str_value
+                    ]
 
-        # TODO: Speed up Ds calculations
-        if im == "Ds595":
-            value = intensity_measures.getDs_nd(DT, accelerations, 5, 95)
+    process_single_value_im("PGV", intensity_measures.get_max_nd, velocities)
+    process_single_value_im("PGA", intensity_measures.get_max_nd, accelerations)
+    process_single_value_im(
+        "CAV", intensity_measures.get_cumulative_abs_velocity_nd, accelerations, times
+    )
+    process_single_value_im(
+        "AI", intensity_measures.get_arias_intensity_nd, accelerations, G, times
+    )
+    process_single_value_im("MMI", intensity_measures.calculate_MMI_nd, velocities)
+    process_single_value_im(
+        "Ds595", intensity_measures.getDs_nd, DT, accelerations, 5, 95
+    )
+    process_single_value_im(
+        "Ds575", intensity_measures.getDs_nd, DT, accelerations, 5, 75
+    )
 
-        if im == "Ds575":
-            value = intensity_measures.getDs_nd(DT, accelerations, 5, 75)
-
-        if im == "AI":
-            value = intensity_measures.get_arias_intensity_nd(accelerations, G, times)
-
-        if im == "CAV":
-            value = intensity_measures.get_cumulative_abs_velocity_nd(
-                accelerations, times
-            )
-
-        if im == "MMI":
-            value = intensity_measures.calculate_MMI_nd(velocities)
-
+    if "pSA" in ims:
+        im = "pSA"
         # store a im type values into a dict {comp: np_array/single float}
         # Geometric is also calculated here
-        value_dict = array_to_dict(value, str_comps, im, comps)
+        psa, spectral_displacements = intensity_measures.get_spectral_acceleration_nd(
+            accelerations, im_options[im], waveform_acc.NT, DT
+        )
+        # Store the pSA im values in the format Tuple(List(periods), dict(component: List(im_values)))
+        # Where the im_values in the component dictionaries correspond to the periods in the periods list
+        pSA_values = array_to_dict(psa, comps_to_calculate, im, comps_to_store)
 
-        # store value dict into the biggest result dict
-        if im == "pSA":
-            result[station_name][im] = (period, value_dict)
+        if {
+            Components.crotd50,
+            Components.crotd100,
+            Components.crotd100_50,
+        }.intersection(comps_to_store):
+            # Only run if any of the given components are selected (Non empty intersection)
+            rotd = calculate_rotd(spectral_displacements, comps_to_store)
+            pSA_values.update(rotd)
+
+        for comp in comps_to_store:
+            if comp.str_value in pSA_values:
+                for i, val in enumerate(im_options[im]):
+                    result[(station_name, comp.str_value)][
+                        f"{im}_{str(val)}"
+                    ] = pSA_values[comp.str_value][i]
+
+    if "FAS" in ims:
+        im = "FAS"
+        try:
+            value = get_fourier_spectrum(accelerations, DT, im_options[im])
+        except FileNotFoundError as e:
+            print(
+                f"Attempting to compute fourier spectrum raised exception: {e}\nThis was most likely caused by attempting to compute for a waveform with more than 16384 timesteps."
+            )
         else:
-            result[station_name][im] = value_dict
+            values_to_store = array_to_dict(
+                value, comps_to_calculate, im, comps_to_store
+            )
+            for comp in comps_to_store:
+                if comp.str_value in values_to_store:
+                    for i, val in enumerate(im_options[im]):
+                        result[(station_name, comp.str_value)][
+                            f"{im}_{str(val)}"
+                        ] = values_to_store[comp.str_value][i]
 
     return result
 
@@ -272,10 +286,10 @@ def compute_measures_multiprocess(
     file_type,
     wave_type,
     station_names,
-    ims=IMS,
+    ims=DEFAULT_IMS,
     comp=None,
-    period=None,
-    output_dir=None,
+    im_options=None,
+    output=None,
     identifier=None,
     rupture=None,
     run_type=None,
@@ -295,8 +309,8 @@ def compute_measures_multiprocess(
     :param station_names:
     :param ims:
     :param comp:
-    :param period:
-    :param output_dir:
+    :param im_options:
+    :param output:
     :param identifier:
     :param rupture:
     :param run_type:
@@ -305,17 +319,24 @@ def compute_measures_multiprocess(
     :param simple_output:
     :return:
     """
-    str_comps_for_int, str_comps = convert_str_comp(comp)
+    (
+        components_to_calculate,
+        components_to_store,
+    ) = constants.Components.get_comps_to_calc_and_store(comp)
+
     bbseries, station_names = get_bbseis(input_path, file_type, station_names)
     total_stations = len(station_names)
-    # determine the size of each iteration base on num of processers
-    steps = get_steps(input_path, process, total_stations)
+    # determine the size of each iteration base on num of processers and mem
+    steps = get_steps(
+        input_path, process, total_stations, "FAS" in ims and bbseries.nt > 32768
+    )
 
     all_result_dict = {}
     p = Pool(process)
+    
     if advanced_im_config.IM_list:
         df_adv_im = {im: pd.DataFrame() for im in advanced_im_config.IM_list}
-
+    
     i = 0
     while i < total_stations:
         # read waveforms of stations
@@ -325,7 +346,7 @@ def compute_measures_multiprocess(
             input_path,
             bbseries,
             stations_to_run,
-            str_comps_for_int,
+            components_to_calculate,
             wave_type=wave_type,
             file_type=file_type,
             units=units,
@@ -334,8 +355,18 @@ def compute_measures_multiprocess(
         array_params = []
         adv_array_params = []
         for waveform in waveforms:
-            array_params.append((waveform, ims, comp, period, str_comps))
-            adv_array_params.append((waveform, advanced_im_config, output_dir))
+            array_params.append(
+                (
+                    waveform,
+                    sorted(ims),
+                    sorted(components_to_store, key=lambda x: x.value),
+                    im_options,
+                    sorted(components_to_calculate, key=lambda x: x.value),
+                )
+            )
+            
+            #array_params.append((waveform, ims, comp, period, str_comps))
+            adv_array_params.append((waveform, advanced_im_config, output))
         # only run simply im if and only if adv_im not going to run
         if not advanced_im_config.IM_list:
             result_list = p.map(compute_measure_single, array_params)
@@ -355,7 +386,7 @@ def compute_measures_multiprocess(
     # write the ouput after all cals are done
     if not advanced_im_config.IM_list:
         write_result(
-            all_result_dict, output_dir, identifier, comp, ims, period, simple_output
+            all_result_dict, output, identifier, simple_output
         )
     # write for advanced IM (pandas array)
     if advanced_im_config.IM_list:
@@ -387,98 +418,34 @@ def get_result_filepath(output_folder, arg_identifier, suffix):
     return os.path.join(output_folder, "{}{}".format(arg_identifier, suffix))
 
 
-def get_header(ims, period):
-    """
-    write header colums for output im_calculations csv file
-    :param ims: a list of im measures
-    :param period: a list of pSA periods
-    :return: header
-    """
-    header = ["station", "component"]
-    psa_names = []
-
-    for im in ims:
-        if im == "pSA":  # only write period if im is pSA.
-            for p in period:
-                if p in BSC_PERIOD:
-                    psa_names.append("pSA_{}".format(p))
-                else:
-                    psa_names.append("pSA_{:.12f}".format(p))
-            header += psa_names
-        else:
-            header.append(im)
-    return header
-
-
-def write_rows(comps, station, ims, result_dict, big_csv_writer, sub_csv_writer=None):
-    """
-    write rows to big csv and, also to single station csvs if not simple output
-    :param comps:
-    :param station:
-    :param ims:
-    :param result_dict:
-    :param big_csv_writer:
-    :param sub_csv_writer:
-    :return: write a single row
-    """
-    for c in comps:
-        row = [station, c]
-        for im in ims:
-            if im != "pSA":
-                row.append(result_dict[station][im][c])
-            else:
-                row += result_dict[station][im][1][c].tolist()
-        big_csv_writer.writerow(row)
-        if sub_csv_writer:
-            sub_csv_writer.writerow(row)
-
-
-def write_result(
-    result_dict, output_folder, identifier, comps, ims, period, simple_output
-):
+def write_result(result_dict, output_folder, identifier, simple_output):
     """
     write a big csv that contains all calculated im value and single station csvs
     :param result_dict:
     :param output_folder:
     :param identifier: user input run name
-    :param comps: a list of comp(s)
-    :param ims: a list of im(s)
-    :param period:
     :param simple_output
     :return:output result into csvs
     """
     output_path = get_result_filepath(output_folder, identifier, ".csv")
-    header = get_header(ims, period)
 
-    # big csv containing all stations
-    with open(output_path, "w") as csv_file:
-        csv_writer = csv.writer(csv_file, delimiter=",", quotechar="|")
-        csv_writer.writerow(header)
-        stations = sorted(result_dict.keys())
+    results_dataframe = pd.DataFrame.from_dict(result_dict, orient="index")
+    results_dataframe.index = pd.MultiIndex.from_tuples(
+        results_dataframe.index, names=["station", "component"]
+    )
+    results_dataframe.sort_values(["station", "component"], inplace=True)
+    results_dataframe = order_im_cols_df(results_dataframe)
 
-        # write each row
-        for station in stations:
-            if not simple_output:  # if single station csvs are needed
-                station_csv = os.path.join(
-                    output_folder,
-                    OUTPUT_SUBFOLDER,
-                    "{}_{}.csv".format(station, "_".join(comps)),
-                )
-                with open(station_csv, "w") as sub_csv_file:
-                    sub_csv_writer = csv.writer(
-                        sub_csv_file, delimiter=",", quotechar="|"
-                    )
-                    sub_csv_writer.writerow(header)
-                    write_rows(
-                        comps,
-                        station,
-                        ims,
-                        result_dict,
-                        csv_writer,
-                        sub_csv_writer=sub_csv_writer,
-                    )
-            else:  # if only the big summary csv is needed
-                write_rows(comps, station, ims, result_dict, csv_writer)
+    # Save the transposed dataframe
+    results_dataframe.to_csv(output_path)
+
+    if not simple_output:
+        # For each subframe with the same station write it to csv
+        for station, sub_frame in results_dataframe.groupby(level=0):
+            station_csv = os.path.join(
+                output_folder, OUTPUT_SUBFOLDER, "{}_{}.csv".format(identifier, station)
+            )
+            sub_frame.to_csv(station_csv)
 
 
 def generate_metadata(output_folder, identifier, rupture, run_type, version):
@@ -500,17 +467,6 @@ def generate_metadata(output_folder, identifier, rupture, run_type, version):
         meta_writer = csv.writer(meta_file, delimiter=",", quotechar="|")
         meta_writer.writerow(["identifier", "rupture", "type", "date", "version"])
         meta_writer.writerow([identifier, rupture, run_type, date, version])
-
-
-def get_im_or_period_help(default_values, im_or_period):
-    """
-    :param default_values: predefined constants
-    :param im_or_period: should be either string "im" or string "period"
-    :return: a help message for input component arg
-    """
-    return "Available and default {}s are: {}".format(
-        im_or_period, ",".join(str(v) for v in default_values)
-    )
 
 
 def validate_input_path(parser, arg_input, arg_file_type):
@@ -538,26 +494,7 @@ def validate_input_path(parser, arg_input, arg_file_type):
             )
 
 
-def validate_im(parser, arg_im):
-    """
-    returns validated user input if pass the validation else raise parser error
-    :param parser:
-    :param arg_im: input
-    :return: validated im(s) in a list
-    """
-    im = arg_im
-    if im != IMS:
-        for m in im:
-            if m not in IMS:
-                parser.error(
-                    "please enter valid im measure name. {}".format(
-                        get_im_or_period_help(IMS, "IM")
-                    )
-                )
-    return im
-
-
-def validate_period(parser, arg_period, arg_extended_period, im):
+def validate_period(arg_period, arg_extended_period):
     """
     returns validated user input if pass the validation else raise parser error
     :param parser:
@@ -569,25 +506,33 @@ def validate_period(parser, arg_period, arg_extended_period, im):
     period = np.array(arg_period, dtype="float64")
 
     if arg_extended_period:
-        period = np.unique(np.append(period, EXT_PERIOD))
-
-    if (arg_extended_period or period.any()) and "pSA" not in im:
-        parser.error(
-            "period or extended period must be used with pSA, but pSA is not in the "
-            "IM measures entered"
-        )
+        period = np.unique(np.append(period, constants.EXT_PERIOD))
 
     return period
 
 
-def get_steps(input_path, nps, total_stations):
+def validate_FAS_frequency(arg_fas_freq):
+    """
+    returns validated user input if pass the validation else raise parser error
+    :param arg_fas_freq: input
+    :return: frequencies in a numpy array
+    """
+    frequencies = np.array(arg_fas_freq, dtype="float64")
+
+    return frequencies
+
+
+def get_steps(input_path, nps, total_stations, high_mem_usage=False):
     """
     :param input_path: user input file/dir path
     :param nps: number of processes
     :param total_stations: total number of stations
+    :param high_mem_usage: If a calculation requiring even larger amounts of RAM is required (ie FAS), this increases the estimated RAM even further
     :return: number of stations per iteration/batch
     """
     estimated_mem = os.stat(input_path).st_size * MEM_FACTOR
+    if high_mem_usage:
+        estimated_mem *= MEM_FACTOR
     available_mem = nps * MEM_PER_CORE
     batches = np.ceil(np.divide(estimated_mem, available_mem))
     steps = int(np.floor(np.divide(total_stations, batches)))